requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
PyYAML>=6.0.1
Pillow>=10.0.0
tqdm>=4.66.0
kaggle>=1.6.14
